+++ src/internal/errors.ts
@@ -151,3 +151,9 @@ export class DuplicateOptionsError extends TypeError {
     super(`Duplicate Option definition at [${duplicateAt.join(',')}] [E032]`);
   }
 }
+
+export class CacheDisabledError extends TypeError {
+  constructor(where: string) {
+    super(`Tried using cache, but was disabled at "${where}" [E033]`);
+  }
+}