+++ src/internal/utils.ts
@@ -21,6 +21,7 @@ import { DecoratorKeys, Severity } from './constants';
 import { constructors, globalOptions } from './data';
 import {
   AssertionFallbackError,
+  CacheDisabledError,
   InvalidOptionsConstructorError,
   NoValidClassError,
   ResolveTypegooseNameError,
@@ -156,6 +157,8 @@ export function getCachedSchema(target: AnyParamConstructor<any>): Record<string
 export function getClass(
   input: mongoose.Document | IObjectWithTypegooseFunction | { typegooseName: string } | string | any
 ): NewableFunction | undefined {
+  assertion(isGlobalCachingEnabled(), () => new CacheDisabledError('getClass'));
+
   if (typeof input === 'string') {
     return constructors.get(input);
   }
@@ -738,4 +741,21 @@ export function mapModelOptionsToNaming(options: IModelOptions | undefined): INa
   }
 
   return mappedNaming;
+}
+
+/**
+ * Helper function to check if caching is enabled globally
+ * @returns "true" if caching is enabled or "false" if disabled
+ */
+export function isGlobalCachingEnabled(): boolean {
+  return !(globalOptions.globalOptions?.disableGlobalCaching === true);
+}
+
+/**
+ * Helper function to check if caching is enabled globally AND by options
+ * @param opt The caching option (from IModelOptions)
+ * @returns "true" if caching is enabled or "false" if disabled
+ */
+export function isCachingEnabled(opt: boolean | undefined): boolean {
+  return isGlobalCachingEnabled() && !(opt === true);
 }